Moved constants and functions related to names and paths from common_storage.ts to names.ts.
