+++ src/reactComponents/ProjectNameComponent.tsx
@@ -22,6 +22,7 @@ import * as Antd from 'antd';
 import * as I18Next from 'react-i18next';
 import * as React from 'react';
 import * as commonStorage from '../storage/common_storage';
+import * as storageNames from '../storage/names';
 
 /** Props for the ProjectNameComponent. */
 interface ProjectNameComponentProps {
@@ -63,7 +64,7 @@ export default function ProjectNameComponent(props: ProjectNameComponentProps): 
       return;
     }
 
-    if (!commonStorage.isValidClassName(newProjectName)) {
+    if (!storageNames.isValidClassName(newProjectName)) {
       showError(newProjectName + INVALID_NAME_MESSAGE_SUFFIX);
       return;
     }
